src/backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.3.0
audioread==3.0.1
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
decorator==5.2.1
distro==1.9.0
ffmpeg-python==0.2.0
filelock==3.13.1
Flask==3.1.0
flask-cors==5.0.1
frozenlist==1.5.0
fsspec==2024.6.1
future==1.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
networkx==3.3
numba==0.61.0
numpy==2.1.3
openai==0.28.0
openai-whisper==20240930
packaging==24.2
platformdirs==4.3.6
pooch==1.8.2
praat-parselmouth==0.4.5
propcache==0.3.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
regex==2024.11.6
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
sentencepiece==0.2.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
srt==3.5.3
sympy==1.13.1
threadpoolctl==3.6.0
tiktoken==0.9.0
torch==2.6.0+cpu
tqdm==4.67.1
triton==3.2.0
typing_extensions==4.12.2
urllib3==2.3.0
vosk==0.3.45
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.18.3
